nyc/assets/ingestion/requirements.txt
pandas>=2.0.0
requests>=2.31.0
python-dateutil>=2.8.0
pyarrow>=14.0.0

